clean up a little more
